AnnapurnaAI/Backend/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.4
psycopg2-binary==2.9.11
Flask-JWT-Extended==4.5.2
python-dotenv==1.0.0
werkzeug==2.3.4
scikit-learn==1.3.0 
pandas==2.0.3
joblib==1.3.2 